+++ src/components/theme/Anontools/Anontools.test.jsx
@@ -11,6 +11,7 @@ describe('Anontools', () => {
   it('renders an anontools component when no token is specified', () => {
     const store = mockStore({
       userSession: { token: null },
+      content: { data: { '@id': 'myid' } },
       intl: {
         locale: 'en',
         messages: {},
@@ -28,6 +29,7 @@ describe('Anontools', () => {
   it('should not render an anontools component when a token is specified', () => {
     const store = mockStore({
       userSession: { token: '1234' },
+      content: { data: {} },
       intl: {
         locale: 'en',
         messages: {},