+++ src/components/theme/Header/Header.test.jsx
@@ -1,14 +1,38 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
+import configureStore from 'redux-mock-store';
+import { Provider } from 'react-intl-redux';
+
 import Header from './Header';
 
-jest.mock('../Logo/Logo', () => jest.fn(() => <div />));
-jest.mock('../SearchWidget/SearchWidget', () => jest.fn(() => <div />));
-jest.mock('../Anontools/Anontools', () => jest.fn(() => <div />));
+const mockStore = configureStore();
+
+jest.mock('../Logo/Logo', () => jest.fn(() => <div id="logo" />));
+jest.mock('../SearchWidget/SearchWidget', () =>
+  jest.fn(() => <div id="searchwidget" />),
+);
+jest.mock('../Anontools/Anontools', () =>
+  jest.fn(() => <div id="anontools" />),
+);
+jest.mock('../Navigation/Navigation', () =>
+  jest.fn(() => <div id="navigation" />),
+);
 
 describe('Header', () => {
   it('renders a header component', () => {
-    const component = renderer.create(<Header pathname="/blog" />);
+    const store = mockStore({
+      userSession: { token: null },
+      intl: {
+        locale: 'en',
+        messages: {},
+      },
+    });
+
+    const component = renderer.create(
+      <Provider store={store}>
+        <Header pathname="/blog" />
+      </Provider>,
+    );
     const json = component.toJSON();
     expect(json).toMatchSnapshot();
   });