+++ CHANGELOG.md
@@ -18,6 +18,14 @@
 * Switched to draft js plugins editor @robgietema
 * Added pastanaga theme @robgietema
 * Added pastanaga manage views @robgietema
+* Added pastanaga theme views @robgietema
+* Added callout styling to draftjs @robgietema
+* Fix paragraph styling in draftjs @robgietema
+* Added image, file and news item view @robgietema
+* Fixed summary and tabular views @robgietema
+* Added social sharing @robgietema
+* Added commenting @robgietema
+* Added tags @robgietema
 
 ## 0.4.0 (2017-05-03)